requirements.txt
pdf2image==1.16.3
python-magic==0.4.27
Pillow>=10.0.0
requests>=2.31.0 